+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- GraphQL clients have method: `transaction(...)` which returns a SuiTransaction for consistency with gRPC clients.
 - [GraphQL beta support](https://github.com/FrankC01/pysui/issues/296)
 - [gRPC beta support](https://github.com/FrankC01/pysui/issues/284)